+++ assembly/src/ast/nodes/advice.rs
@@ -18,7 +18,8 @@ pub enum AdviceInjectorNode {
     PushU64div,
     PushExt2intt,
     PushSmtGet,
-    PushSmtInsert,
+    PushSmtSet,
+    PushSmtPeek,
     PushMapVal,
     PushMapValImm { offset: u8 },
     PushMapValN,
@@ -37,7 +38,8 @@ impl From<&AdviceInjectorNode> for AdviceInjector {
             PushU64div => Self::DivU64,
             PushExt2intt => Self::Ext2Intt,
             PushSmtGet => Self::SmtGet,
-            PushSmtInsert => Self::SmtInsert,
+            PushSmtSet => Self::SmtSet,
+            PushSmtPeek => Self::SmtPeek,
             PushMapVal => Self::MapValueToStack {
                 include_len: false,
                 key_offset: 0,
@@ -72,7 +74,8 @@ impl fmt::Display for AdviceInjectorNode {
             PushU64div => write!(f, "push_u64div"),
             PushExt2intt => write!(f, "push_ext2intt"),
             PushSmtGet => write!(f, "push_smtget"),
-            PushSmtInsert => write!(f, "push_smtinsert"),
+            PushSmtSet => write!(f, "push_smtset"),
+            PushSmtPeek => write!(f, "push_smtpeek"),
             PushMapVal => write!(f, "push_mapval"),
             PushMapValImm { offset } => write!(f, "push_mapval.{offset}"),
             PushMapValN => write!(f, "push_mapvaln"),
@@ -92,16 +95,17 @@ impl fmt::Display for AdviceInjectorNode {
 const PUSH_U64DIV: u8 = 0;
 const PUSH_EXT2INTT: u8 = 1;
 const PUSH_SMTGET: u8 = 2;
-const PUSH_SMTINSERT: u8 = 3;
-const PUSH_MAPVAL: u8 = 4;
-const PUSH_MAPVAL_IMM: u8 = 5;
-const PUSH_MAPVALN: u8 = 6;
-const PUSH_MAPVALN_IMM: u8 = 7;
-const PUSH_MTNODE: u8 = 8;
-const INSERT_MEM: u8 = 9;
-const INSERT_HDWORD: u8 = 10;
-const INSERT_HDWORD_IMM: u8 = 11;
-const INSERT_HPERM: u8 = 12;
+const PUSH_SMTSET: u8 = 3;
+const PUSH_SMTPEEK: u8 = 4;
+const PUSH_MAPVAL: u8 = 5;
+const PUSH_MAPVAL_IMM: u8 = 6;
+const PUSH_MAPVALN: u8 = 7;
+const PUSH_MAPVALN_IMM: u8 = 8;
+const PUSH_MTNODE: u8 = 9;
+const INSERT_MEM: u8 = 10;
+const INSERT_HDWORD: u8 = 11;
+const INSERT_HDWORD_IMM: u8 = 12;
+const INSERT_HPERM: u8 = 13;
 
 impl Serializable for AdviceInjectorNode {
     fn write_into<W: ByteWriter>(&self, target: &mut W) {
@@ -110,7 +114,8 @@ impl Serializable for AdviceInjectorNode {
             PushU64div => target.write_u8(PUSH_U64DIV),
             PushExt2intt => target.write_u8(PUSH_EXT2INTT),
             PushSmtGet => target.write_u8(PUSH_SMTGET),
-            PushSmtInsert => target.write_u8(PUSH_SMTINSERT),
+            PushSmtSet => target.write_u8(PUSH_SMTSET),
+            PushSmtPeek => target.write_u8(PUSH_SMTPEEK),
             PushMapVal => target.write_u8(PUSH_MAPVAL),
             PushMapValImm { offset } => {
                 target.write_u8(PUSH_MAPVAL_IMM);
@@ -139,7 +144,8 @@ impl Deserializable for AdviceInjectorNode {
             PUSH_U64DIV => Ok(AdviceInjectorNode::PushU64div),
             PUSH_EXT2INTT => Ok(AdviceInjectorNode::PushExt2intt),
             PUSH_SMTGET => Ok(AdviceInjectorNode::PushSmtGet),
-            PUSH_SMTINSERT => Ok(AdviceInjectorNode::PushSmtInsert),
+            PUSH_SMTSET => Ok(AdviceInjectorNode::PushSmtSet),
+            PUSH_SMTPEEK => Ok(AdviceInjectorNode::PushSmtPeek),
             PUSH_MAPVAL => Ok(AdviceInjectorNode::PushMapVal),
             PUSH_MAPVAL_IMM => {
                 let offset = source.read_u8()?;